pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "django>=5.1.7",
    "django-cors-headers>=4.7.0",
    "djangorestframework>=3.16.0",
    "numpy>=2.2.4",
    "opencv-python>=4.11.0.86",
    "pillow>=11.1.0",
    "python-dotenv>=1.1.0",
]
